test/web-platform-tests/to-upstream/html/semantics/forms/the-input-element/disabled-checkbox.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- TODO: fix below -->
<title>not sure what to name this or where to put it</title>
<!-- TODO: fix below -->
<link rel=help href="https://dom.spec.whatwg.org/#dom-element-hasattribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!-- TODO: find out what this is -->
<span data-e2="2" data-F2="3" id="t"></span>

<script>
"use strict";

test(() => {

  const input = document.createElement("input");
  input.type = "checkbox";
  input.disabled = true;
  document.body.appendChild(input);
  const events = [];

  input.addEventListener("change", () => { events.push("change"); });
  input.addEventListener("click", () => { events.push("click"); });
  input.addEventListener("input", () => { events.push("input"); });

  // disabled checkbox begins unchecked
  assert_true(!input.checked);

  input.click();

  // disabled checkbox still unchecked after .click()
  assert_true(!input.checked);

  // disabled checkbox emits no events after synthetic .click()
  assert_array_equals(events, []);

  const event = new MouseEvent("click", { bubbles: true, cancelable: true });
  input.dispatchEvent(event);
  // disabled checkbox is checked after dispatching direct click event
  assert_true(input.checked);
  // disabled checkbox runs click activation events after dispatchEvent
  assert_array_equals(events, ["click", "input", "change"]);

}, "checkbox input respects disabled attribute for synthetic events but not for dispatchEvent");
</script>
